+++ courses/new-password.php
@@ -19,29 +19,32 @@
 //  ********** REST PASSWORD ************//
  if(Input::post('reset_password'))
  {
-     $validate = new DB(); 
-     $validation = $validate->validate([
-            'new_password' => 'required|min:6|max:12',
-            'confirm_password' => 'required|min:6|max:12|match:new_password'
-     ]);
-
-     if(!$validation->passed())
+    if(Token::check())
     {
-        return back();
-    }
+        $validate = new DB(); 
+        $validation = $validate->validate([
+                'new_password' => 'required|min:6|max:12',
+                'confirm_password' => 'required|min:6|max:12|match:new_password'
+        ]);
 
-    $reset = $connection->select('course_user_reset_password')->where('reset_token', Input::get('reset_token'))->first();
-    if($reset)
-    {
-        $update_password = $connection->update('course_users', [
-            'password' => password_hash(Input::get('new_password'), PASSWORD_DEFAULT)
-        ])->where('email', $reset->reset_email)->save();
-
-        if($update_password)
+        if(!$validation->passed())
         {
-            $connection->delete('course_user_reset_password')->where('reset_token', Input::get('reset_token'))->save();
-            Session::flash('success', 'Password reset successfully, you can login!');
-            return view('/courses/login');
+            return back();
+        }
+
+        $reset = $connection->select('course_user_reset_password')->where('reset_token', Input::get('reset_token'))->first();
+        if($reset)
+        {
+            $update_password = $connection->update('course_users', [
+                'password' => password_hash(Input::get('new_password'), PASSWORD_DEFAULT)
+            ])->where('email', $reset->reset_email)->save();
+
+            if($update_password)
+            {
+                $connection->delete('course_user_reset_password')->where('reset_token', Input::get('reset_token'))->save();
+                Session::flash('success', 'Password reset successfully, you can login!');
+                return view('/courses/login');
+            }
         }
     }
  }
@@ -113,6 +116,7 @@
                                           </div>
                                      </div>
                                 </div>
+                                <?= csrf_token() ?>
 							</form>
 						</div>
                     </div>
