+++ employee/login.php
@@ -9,64 +9,66 @@
 
 
 
-//==================================================
-// EMAIL LOGIN
-// =================================================
+
+
+// ***********  EMAIL LOGIN ************//
  if(Input::post('login_employee'))
  {
-    $validate = new DB();
-    
-    $validation = $validate->validate([
-        'email' => 'required|email',
-        'password' => 'required|min:6|max:12',
-    ]);
-
-    if(!$validation->passed())
+    if(Token::check())
     {
-        return back();
-    }
-      
-    if($validation->passed())
-    {
-        $verify = new DB();
-        $verification = $verify->select('employee')->where('email', Input::get('email'))->first();
-        if(!$verification)
+        $validate = new DB();
+        
+        $validation = $validate->validate([
+            'email' => 'required|email',
+            'password' => 'required|min:6|max:12',
+        ]);
+
+        if(!$validation->passed())
         {
-            Session::errors('errors', ['email' => '*Wrong email provided, try again!']);
             return back();
         }
-
-        if(!password_verify(Input::get('password'), $verification->password))
+        
+        if($validation->passed())
         {
-            Session::errors('errors', ['password' => '*Wrong password, try again!']);
-            return back();
-        }
+            $verify = new DB();
+            $verification = $verify->select('employee')->where('email', Input::get('email'))->first();
+            if(!$verification)
+            {
+                Session::errors('errors', ['email' => '*Wrong email provided, try again!']);
+                return back();
+            }
 
-        if($verification->e_is_deactivate)
-        {
-            Session::flash('error', '*This account has been deactivated, please contact the admin.');
-            return back();
-        }
+            if(!password_verify(Input::get('password'), $verification->password))
+            {
+                Session::errors('errors', ['password' => '*Wrong password, try again!']);
+                return back();
+            }
 
-        $remember_me = Input::get('remember_me')? true : false;
+            if($verification->e_is_deactivate)
+            {
+                Session::flash('error', '*This account has been deactivated, please contact the admin.');
+                return back();
+            }
 
-        $logged_in = Auth_employee::login(Input::get('email'), $remember_me);
+            $remember_me = Input::get('remember_me')? true : false;
 
-        if($logged_in && Session::has('old_url'))
-        {
-            $old_url = Session::get('old_url');
-            Session::delete('old_url');
-            Session::flash('success', 'You have loggedin successfully!');
-            return view($old_url);
-        }
+            $logged_in = Auth_employee::login(Input::get('email'), $remember_me);
+
+            if($logged_in && Session::has('old_url'))
+            {
+                $old_url = Session::get('old_url');
+                Session::delete('old_url');
+                Session::flash('success', 'You have loggedin successfully!');
+                return view($old_url);
+            }
 
             if($logged_in)
             {
                 Session::flash('success', 'You have loggedin successfully!');
                 return view('/');
             }
+        }
     }
-    
 }
 
 
@@ -176,6 +178,7 @@ if(Input::post('facebook_login'))
                             <div class="col-lg mb-3">
                                 <button type="submit" name="google_login" class="btn btn-block color-white bgc-gogle"><i class="fa fa-google float-left  mb-2"></i> Google</button>
                             </div>
+                            <?= csrf_token() ?>
                         </div>
                     </div>
                 </div>
